requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
gunicorn==20.1.0
requests==2.26.0
pyotp==2.6.0
qrcode[pil]==7.3.1
djangorestframework-simplejwt==5.2.0